gitignore: Ignore more build system products
Ignore static libraries, cmake products, generated code and unit test
binaries.

Some extra care must be taken not to ignore more than intended when ignoring
the *_UnitTests. There is no way to directly express a path-neutral "only
ignore files with this pattern, not directories" in a single gitignore
pattern, but it can be expressed as "ignore everything containing this pattern
except directories".