Complete instantiated types
RivReservoirPartMgr uses a Collection of RivGridPartMgr. The Collection
indirectly causes the release() method of the Object base class of
RivGridPartMgr to be called as part of its destructor, but this is not
allowed for incomplete types and RivGridPartMgr is only forward-declared
at the time of the instantiation (in RivReservoirPartMgr).

A similar problem occurs in the use of RigStatisticsDataCache in the
type RigCaseCellResultsData.